Modified settings to allow email sending
