+++ smse_onboarding_main_project/smse_onboarding/settings.py
@@ -12,6 +12,10 @@ https://docs.djangoproject.com/en/5.1/ref/settings/
 
 from pathlib import Path
 import os
+import certifi
+import ssl
+import urllib3
+from django.core.mail.backends.smtp import EmailBackend
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
@@ -93,13 +97,11 @@ TEMPLATES = [
 ]
 
 CSRF_TRUSTED_ORIGINS = [
-    # 'https://smse-onboarding.dedyn.io:8080',
-    # 'https://smse-onboarding.dedyn.io:8000',
     'https://smse-onboarding.dedyn.io',
-    # 'http://smse-onboarding.dedyn.io:8080',
-    # 'http://smse-onboarding.dedyn.io:8000',
     'http://smse-onboarding.dedyn.io',
     'http://localhost:8000',
+    'http://127.0.0.1:8000',
+    'http://0.0.0.0:8000',
 ]
 
 CSRF_COOKIE_DOMAIN = 'smse-onboarding.dedyn.io'
@@ -202,15 +204,6 @@ FIXTURE_DIRS = [
 
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
-# Mail configuration
-EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
-EMAIL_HOST = 'smtp.gmail.com'
-EMAIL_PORT = 587
-EMAIL_USE_TLS = True
-EMAIL_HOST_USER = 'usdsmse@gmail.com'
-EMAIL_HOST_PASSWORD = 'avslwlzrooptdlbm'
-DEFAULT_FROM_EMAIL = 'usdsmse@gmail.com'
-
 LOGIN_REDIRECT_URL = '/home/'
 LOGOUT_REDIRECT_URL = '/users/login/'
 LOGIN_URL = '/users/login/'
@@ -262,3 +255,52 @@ LOGGING = {
         },
     },
 }
+
+# Disable SSL verification warnings
+urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
+
+# Custom email backend that ignores SSL verification
+class CustomEmailBackend(EmailBackend):
+    def open(self):
+        if self.connection:
+            return False
+        try:
+            # Create unverified SSL context
+            context = ssl._create_unverified_context()
+            
+            # Create SMTP connection
+            self.connection = self.connection_class(
+                self.host, self.port,
+                timeout=self.timeout
+            )
+            
+            # Start TLS with unverified context
+            if self.use_tls:
+                self.connection.starttls(context=context)
+            
+            # Login if credentials are provided
+            if self.username and self.password:
+                self.connection.login(self.username, self.password)
+            return True
+        except Exception:
+            if not self.fail_silently:
+                raise
+            return False
+
+# Override email settings for development
+EMAIL_BACKEND = 'smse_onboarding.settings_dev.CustomEmailBackend'
+EMAIL_HOST = 'smtp.gmail.com'
+EMAIL_PORT = 587
+EMAIL_USE_TLS = True
+EMAIL_HOST_USER = 'usdsmse@gmail.com'
+EMAIL_HOST_PASSWORD = 'avslwlzrooptdlbm'
+DEFAULT_FROM_EMAIL = 'usdsmse@gmail.com'
+
+# Configure SSL settings
+os.environ['SSL_CERT_FILE'] = certifi.where()
+os.environ['REQUESTS_CA_BUNDLE'] = certifi.where()
+os.environ['CURL_CA_BUNDLE'] = certifi.where()
+os.environ['PYTHONHTTPSVERIFY'] = '0'
+
+# Configure SSL context
+ssl._create_default_https_context = ssl._create_unverified_context
